app/aurion-hosted-agent/pyproject.toml
[project]
name = "aurion-hosted-agent"
version = "0.1.0"
description = "Aurion Claim Intake — simplified hosted agent reference"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agent-framework==1.0.0rc3",
    "agent-framework-devui==1.0.0b260304",
    "azure-ai-agentserver-core==1.0.0b16",
    "azure-ai-agentserver-agentframework==1.0.0b16",
    "azure-identity>=1.25.2",
    "python-dotenv>=1.2.2",
    "pydantic>=2.0,<2.13",
]

[tool.uv]
prerelease = "allow"
